+++ tests/validator/test_workspace_validator.py
@@ -202,6 +202,10 @@ class TestWorkspaceValidator(TestCase):
         )
         self.assertEqual(len(report.errors), 0)
 
+    def test_pcgtsid(self):
+        report = WorkspaceValidator.validate(self.resolver, assets.path_to('kant_aufklaerung_1784/data/mets.xml'))
+        self.assertIn('pc:PcGts/@pcGtsId differs from mets:file/@ID: "" !== "PAGE_0017_PAGE"', report.warnings)
+
 
 if __name__ == '__main__':
     main()