+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Versioned according to [Semantic Versioning](http://semver.org/).
 Changed:
 
   * `OcrdFile` constructor accepts `ID` parameter
+  * `model_factory.page_from_image` now sets the `@pcGtsId` attribute tot the file's `@ID`, #378
+  * `WorkspaceValidator`:  check `pc:PcGts/@pcGtsId` == `mets:file/@ID`, #378
   * `OcrdFile` constructor: removed long-obsolete `instance` parameter
   * METS: reorder elements according to schema in empty METS, #487